+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Flask command to delete models from m_l_component collection [@Michal-Kolomanski]
 - Flask train, db command and docs in the README [@Michal-Kolomanski]
 - If requested recommendation model (in /recommendation body request) does not exist try to use another model [@Michal-Kolomanski]
+- engine_version in recommendation collection [@Michal-Kolomanski]
 ### Changed
 - Black version = 20.8b0 - click dependency [@Michal-Kolomanski]
 ### Fixed